refactor(playground-ui): simplify AutoBePlaygroundChatPromptMovie and AutoBePlaygroundChatUploadMovie components
- Removed unused state and functions from AutoBePlaygroundChatPromptMovie to streamline the component.
- Enhanced AutoBePlaygroundChatUploadMovie by integrating file upload functionality and improving user interaction with file selection and messaging.
- Updated UI elements for better user experience and code clarity.
